feat(solace-message-client): provide support for zoneless change detection
