+++ apps/solace-message-client-testing-app/src/app/app.config.ts
@@ -1,4 +1,4 @@
-import {ApplicationConfig, EnvironmentProviders, makeEnvironmentProviders, provideZoneChangeDetection} from '@angular/core';
+import {ApplicationConfig, EnvironmentProviders, makeEnvironmentProviders} from '@angular/core';
 import {provideRouter, withHashLocation} from '@angular/router';
 import {routes} from './app.routes';
 import {MAT_FORM_FIELD_DEFAULT_OPTIONS} from '@angular/material/form-field';
@@ -12,7 +12,6 @@ export const appConfig: ApplicationConfig = {
     provideRouter(routes, withHashLocation()),
     provideAnimations(),
     provideMaterialDefaults(),
-    provideZoneChangeDetection(),
   ],
 };
