+++ docs/site/getting-started.md
@@ -7,9 +7,6 @@
 
 This short manual explains how to install and use the library.
 
-> [!WARNING]
-> `@solace-community/angular-solace-message-client` does not support zoneless mode and may not work as expected. Support is planned for 2026.
-
 <details>
   <summary><strong>Step 1: Install the Library from NPM</strong></summary>
   <br>
@@ -29,7 +26,7 @@ npm install @solace-community/angular-solace-message-client solclientjs @scion/t
 Open `app.config.ts` and register Solace Message Client providers.
 
 ```ts
-import {ApplicationConfig, provideZoneChangeDetection} from '@angular/core';
+import {ApplicationConfig} from '@angular/core';
 import {provideSolaceMessageClient} from '@solace-community/angular-solace-message-client';
 
 export const appConfig: ApplicationConfig = {
@@ -40,7 +37,6 @@ export const appConfig: ApplicationConfig = {
       userName: 'YOUR USERNAME',
       password: 'YOUR PASSWORD',
     }),
-    provideZoneChangeDetection(), // provide zone change detection until `@solace-community/angular-solace-message-client` supports zoneless
   ],
 };
 ```
@@ -50,7 +46,6 @@ If you are not using `app.config.ts`, register the Solace Message Client directl
 ```ts
 import {bootstrapApplication} from '@angular/platform-browser';
 import {provideSolaceMessageClient} from '@solace-community/angular-solace-message-client';
-import {provideZoneChangeDetection} from '@angular/core';
 
 bootstrapApplication(AppComponent, {
   providers: [
@@ -60,7 +55,6 @@ bootstrapApplication(AppComponent, {
       userName: 'YOUR USERNAME',
       password: 'YOUR PASSWORD',
     }),
-    provideZoneChangeDetection(), // provide zone change detection until `@solace-community/angular-solace-message-client` supports zoneless
   ],
 });
 ```
@@ -69,7 +63,7 @@ Or for `NgModule` based applications:
 
 ```ts
 import {provideSolaceMessageClient} from '@solace-community/angular-solace-message-client';
-import {NgModule, provideZoneChangeDetection} from '@angular/core';
+import {NgModule} from '@angular/core';
 
 @NgModule({
   providers: [
@@ -79,7 +73,6 @@ import {NgModule, provideZoneChangeDetection} from '@angular/core';
       userName: 'YOUR USERNAME',
       password: 'YOUR PASSWORD',
     }),
-    provideZoneChangeDetection(), // provide zone change detection until `@solace-community/angular-solace-message-client` supports zoneless
   ],
   bootstrap: [
     AppComponent,
