+++ projects/solace-message-client/src/lib/solace-message-client-browse.spec.ts
@@ -1,7 +1,6 @@
 import {MessageEnvelope, SolaceMessageClient} from './solace-message-client';
 import {ObserveCaptor} from '@scion/toolkit/testing';
 import {TestBed} from '@angular/core/testing';
-import {NgZone} from '@angular/core';
 import {QueueBrowserEventName, QueueBrowserProperties, QueueDescriptor, QueueType, SDTFieldType, SDTMapContainer, SessionEventCode} from 'solclientjs';
 import {SessionFixture} from './testing/session.fixture';
 import {createOperationError, createQueueMessage, drainMicrotaskQueue, initSolclientFactory} from './testing/testing.utils';
@@ -107,58 +106,6 @@ describe('SolaceMessageClient - Browse', () => {
       jasmine.objectContaining<MessageEnvelope>({message: msg2}),
       jasmine.objectContaining<MessageEnvelope>({message: msg3}),
     ]);
-  });
-
-  it('should receive messages in the zone subscribed (inside Angular)', async () => {
-    const sessionFixture = new SessionFixture();
-    const queueBrowserFixture = sessionFixture.queueBrowserFixture;
-    const observeCaptor = new ObserveCaptor(() => NgZone.isInAngularZone());
-    TestBed.configureTestingModule({
-      providers: [
-        provideSolaceMessageClient({url: 'url', vpnName: 'vpn'}),
-        provideSession(sessionFixture),
-      ],
-    });
-    const solaceMessageClient = TestBed.inject(SolaceMessageClient);
-    await sessionFixture.simulateEvent(SessionEventCode.UP_NOTICE);
-
-    // Subscribe inside the Angular zone.
-    TestBed.inject(NgZone).run(() => solaceMessageClient.browse$('queue').subscribe(observeCaptor));
-
-    // Simulate the queue browser to be connected to the broker
-    await queueBrowserFixture.simulateEvent(QueueBrowserEventName.UP);
-
-    // Simulate to receive a message
-    await queueBrowserFixture.simulateMessage(createQueueMessage('queue'));
-
-    // Expect message to be received inside the Angular zone
-    expect(observeCaptor.getValues()).toEqual([true]);
-  });
-
-  it('should receive messages in the zone subscribed (outside Angular)', async () => {
-    const sessionFixture = new SessionFixture();
-    const queueBrowserFixture = sessionFixture.queueBrowserFixture;
-    const observeCaptor = new ObserveCaptor(() => NgZone.isInAngularZone());
-    TestBed.configureTestingModule({
-      providers: [
-        provideSolaceMessageClient({url: 'url', vpnName: 'vpn'}),
-        provideSession(sessionFixture),
-      ],
-    });
-    const solaceMessageClient = TestBed.inject(SolaceMessageClient);
-    await sessionFixture.simulateEvent(SessionEventCode.UP_NOTICE);
-
-    // Subscribe outside the Angular zone.
-    TestBed.inject(NgZone).runOutsideAngular(() => solaceMessageClient.browse$('queue').subscribe(observeCaptor));
-
-    // Simulate the queue browser to be connected to the broker
-    await queueBrowserFixture.simulateEvent(QueueBrowserEventName.UP);
-
-    // Simulate to receive a message
-    await queueBrowserFixture.simulateMessage(createQueueMessage('queue'));
-
-    // Expect message to be received outside the Angular zone
-    expect(observeCaptor.getValues()).toEqual([false]);
   });
 
   it('should start the queue browser when connected to the broker (UP)', async () => {
