+++ projects/solace-message-client/src/lib/solace-message-client-consume.spec.ts
@@ -1,7 +1,6 @@
 import {MessageEnvelope, SolaceMessageClient} from './solace-message-client';
 import {ObserveCaptor} from '@scion/toolkit/testing';
 import {TestBed} from '@angular/core/testing';
-import {NgZone} from '@angular/core';
 import {MessageConsumerEventName, MessageConsumerProperties, QueueDescriptor, QueueType, SDTFieldType, SDTMapContainer, SessionEventCode, SolclientFactory} from 'solclientjs';
 import {SessionFixture} from './testing/session.fixture';
 import {createOperationError, createTopicMessage, drainMicrotaskQueue, initSolclientFactory} from './testing/testing.utils';
@@ -143,58 +142,6 @@ describe('SolaceMessageClient - Consume', () => {
       jasmine.objectContaining<MessageEnvelope>({message: message1}),
       jasmine.objectContaining<MessageEnvelope>({message: message2}),
     ]);
-  });
-
-  it('should receive messages in the zone subscribed (inside Angular)', async () => {
-    const sessionFixture = new SessionFixture();
-    const messageConsumerFixture = sessionFixture.messageConsumerFixture;
-    const observeCaptor = new ObserveCaptor(() => NgZone.isInAngularZone());
-    TestBed.configureTestingModule({
-      providers: [
-        provideSolaceMessageClient({url: 'url', vpnName: 'vpn'}),
-        provideSession(sessionFixture),
-      ],
-    });
-    const solaceMessageClient = TestBed.inject(SolaceMessageClient);
-    await sessionFixture.simulateEvent(SessionEventCode.UP_NOTICE);
-
-    // Subscribe inside the Angular zone.
-    TestBed.inject(NgZone).run(() => solaceMessageClient.consume$('topic').subscribe(observeCaptor));
-
-    // Simulate the message consumer to be connected to the broker
-    await messageConsumerFixture.simulateEvent(MessageConsumerEventName.UP);
-
-    // Simulate to receive a message
-    await messageConsumerFixture.simulateMessage(createTopicMessage('topic'));
-
-    // Expect message to be received inside the Angular zone
-    expect(observeCaptor.getValues()).toEqual([true]);
-  });
-
-  it('should receive messages in the zone subscribed (outside Angular)', async () => {
-    const sessionFixture = new SessionFixture();
-    const messageConsumerFixture = sessionFixture.messageConsumerFixture;
-    const observeCaptor = new ObserveCaptor(() => NgZone.isInAngularZone());
-    TestBed.configureTestingModule({
-      providers: [
-        provideSolaceMessageClient({url: 'url', vpnName: 'vpn'}),
-        provideSession(sessionFixture),
-      ],
-    });
-    const solaceMessageClient = TestBed.inject(SolaceMessageClient);
-    await sessionFixture.simulateEvent(SessionEventCode.UP_NOTICE);
-
-    // Subscribe outside the Angular zone.
-    TestBed.inject(NgZone).runOutsideAngular(() => solaceMessageClient.consume$('topic').subscribe(observeCaptor));
-
-    // Simulate the message consumer to be connected to the broker
-    await messageConsumerFixture.simulateEvent(MessageConsumerEventName.UP);
-
-    // Simulate to receive a message
-    await messageConsumerFixture.simulateMessage(createTopicMessage('topic'));
-
-    // Expect message to be received outside the Angular zone
-    expect(observeCaptor.getValues()).toEqual([false]);
   });
 
   it('should error on connection error (CONNECT_FAILED_ERROR)', async () => {
